apps/api_v2/requirements.txt
diskcache==5.6.3
fastapi==0.112.3
numpy==2.1.1
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==1.4.52
uvicorn==0.30.6
